+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 dd mm yyyy
 
 - New: The user agent for the HTTP client is now configurable (optional).
+- New: Configuration to reduce ESI 403 errors from structure name updates (Administration -> Settings -> 
+  Features -> Structure Name Updates).
 - New: The time of the last check was added to the ESI tokens.
 - New: Added "characters" option to the `check-token` command to optionally reduce the number of ESI tokens that
   are validated.
 - Change: ESI tokens that are marked as invalid are no longer validated.
-- Fix: Added missing director token for structure name updates (member tracking).
+- Fix: Re-added missing director token for structure name updates (member tracking).
 - Text improvements.
 - Updated dependencies.
 - Dev: Removed Node.js 14 support.